Rewrite as Loopscooper: beat-locked detection, web UI overhaul
- Rename project pymusiclooper -> loopscooper (package, CLI entry point,
  docs); single-command CLI plus a Flask web UI
- Anchor loop candidates to actually-detected beats an exact beat-count
  apart instead of a duration window, so loops land on the beat instead of
  merely being close to the right length; a damped phase correction keeps
  the beat grid from locking onto syncopated hits or a doubled pulse
- Target 2-bar and 4-bar loops by default; surface 1-bar loops only at
  a 90%+ match score
- Web UI: retro Akai/E-mu-style theme (beige chassis, grey control panel,
  red accents), dot-matrix LCD readout for filename/duration/BPM/key,
  loop candidates paginated 8/page with selections persisting across
  pages, eject button to load a new file without reloading
- Fix web preview using the AudioContext's (possibly resampled) sample
  rate instead of the source file's actual rate to convert loop points,
  which silently scaled loop length in the browser even though exported
  WAVs were correct
- Fix overlapping/unstoppable preview playback when a candidate was
  clicked twice before the previous decode finished

+++ loopscooper/audio.py
@@ -3,13 +3,13 @@ import os
 import librosa
 import numpy as np
 
-from pymusiclooper.exceptions import AudioLoadError
+from loopscooper.exceptions import AudioLoadError
 
 
 class MLAudio:
     """Wrapper class for loading audio files and containing the necessary audio data for PyMusicLooper."""
 
-    total_duration: int
+    total_duration: float
     filepath: str
     filename: str
     audio: np.ndarray
@@ -97,7 +97,7 @@ class MLAudio:
     def frames_to_ftime(self, frame: int):
         time_sec = librosa.core.frames_to_time(frame, sr=self.rate)
         return f"{time_sec // 60:02.0f}:{time_sec % 60:06.3f}"
-    
+
     def samples_to_ftime(self, samples: int):
         time_sec = librosa.core.samples_to_time(samples, sr=self.rate)
         return f"{time_sec // 60:02.0f}:{time_sec % 60:06.3f}"
